backend/requirements.txt
# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0

# 数据库相关
sqlalchemy==1.4.53
aiomysql==0.2.0
PyMySQL==1.1.0
mysqlclient==2.2.0
databases[mysql]==0.8.0
alembic==1.13.1

# 异步HTTP客户端
httpx==0.25.2
aiofiles==23.2.1

# 认证和安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.6

# 数据验证
pydantic==2.5.0
pydantic-settings==2.1.0

# 任务队列
celery==5.3.4
redis==5.0.1

# 日志
structlog==23.2.0

# 工具库
python-dotenv==1.0.0
email-validator==2.1.0
Jinja2==3.1.2

# 大模型相关
openai==1.6.1
dashscope==1.19.0
google-generativeai==0.3.2
tenacity==8.2.3

# 文件处理
python-docx==1.1.0
openpyxl==3.1.2
markdown==3.5.1
beautifulsoup4==4.12.2
lxml==4.9.3

# 开发依赖
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1 